Checked what happens when trying to create a copy of an object. I was only able to copy the reference to the object's instance and any changes were reflected in the copy too. This was when using:
hero1= new Hero();
hero2 = new Hero();
hero2= hero1;

I'll have to be careful when creating my code and not accidentally delete an instance's reference by overwriting it with another instance's reference when trying to duplicate an instance of an object. I'll think carefully about how my constructors will work and handle the loading of duplicate [eg. Monsters] and just have the constructor do all the copying before an object is fully defined and independent.

Oh yes one more thing, as far as I know these sort of rules only apply to user defined classes (eg. Unit or Monster). datatypes like int and string should be able to be copied and have a separate instance by using similar methods to c;

int x=30;
int y=x;
x=35;
Console.Writeline(y);

should say 30, because the y is a separate instance of an int and the line y=x only sets the value of y equal to the value of x, where as for a user defined class it would set x and y to references for the same instance of that object. Also I think I shall try this out next just to confirm.